pyproject.toml
[tool.poetry]
name = "jaerimlee"
version = "0.1.0"
description = ""
authors = ["leejaerim <ger6608@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"